refactor: consolidate imports from src/index.js across components
