+++ src/index.js
@@ -1,0 +1,8 @@
+import html from "./html/index.js";
+export * from "./html/custom/fragment.js";
+export * from "./html/native.js";
+export * from "./core/reference.js";
+export * from "./core/conditional.js";
+export * from "./router/index.js";
+export * from "./router/Link.js";
+export default html;
